Add signature test cases
The existing ones changed to take the BackwardCompatible inference mode.
All Json and Xml cases that had value inference enabled have been copied
to 3 new cases: one for each value of InferenceMode.

The expected content is always the same: enabling inline schemas
when no schema is defined in the source file should not change anything to the output.